made cache util
